+++ changelog.md
@@ -1,5 +1,8 @@
 # changelog
 
+## 2.0.4
+* `FIX` [#550](https://github.com/sumneko/lua-language-server/issues/550)
+
 ## 2.0.3
 `2021-6-24`
 * `CHG` improve memory usage